Move `SourceHighlighter` to `CodeDumper`
